// src/interfaces/product.interface.ts
export interface ProductInterface {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: Rating;
}

export interface Rating {
  rate: number;
  count: number;
}

export interface ProductInCart {
  product: ProductInterface;
  quantity: number;
}